HS-102 FEAT: is_builtin   관련 함수 분리
